小Bug修改

+++ Gamesssss/Gamesssss/Assets/Scripts/KWUtility.cs
@@ -4,8 +4,8 @@ using System;
 
 public class KWUtility  {
 
-    public static int Random( int min, int max ) {
-        int result = UnityEngine.Random.Range( min, max );
+    public static int Random( float min, float max ) {
+        int result = UnityEngine.Random.Range( (int) min, (int)max );
         return result;
     }
